Tests/LibWeb/Text/expected/navigator-storage-estimate.txt

usage is a number: true
quota is a number: true
usage does not exceed quota: true
usage reflects written localStorage: true
